requirements.txt
openai==0.28
streamlit
openai
langchain
beautifulsoup4
requests
python-dotenv
